requirements.txt
Django>=3.0,<4.0
djangorestframework>=3.11.0
mysqlclient>=2.0.3
